# packages/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Plinth API
  version: 1.0.0
  description: |
    Production-grade multi-tenant SaaS API with comprehensive security controls.

    ## Rate Limiting

    This API implements per-endpoint rate limiting to protect against abuse and ensure fair usage:

    ### Authentication Endpoints (Most Restrictive)
    - **Login**: 5 requests per minute
    - **Register**: 3 requests per hour
    - **Password Change**: 3 requests per 15 minutes
    - **Refresh Token**: 20 requests per 15 minutes

    ### Invitation Operations (Spam Prevention)
    - **Create Invitation**: 20 requests per hour
    - **Accept Invitation**: 10 requests per 15 minutes

    ### API Key Operations
    - **Create API Key**: 10 requests per hour

    ### Generic Operations
    - **Read Operations (GET)**: 300 requests per 15 minutes
    - **Write Operations (POST/PATCH/DELETE)**: 100 requests per 15 minutes

    ### Global Fallback
    - **All Other Endpoints**: 1000 requests per 15 minutes

    When rate limit is exceeded, the API returns HTTP 429 with standard rate limiting headers including `Retry-After`.
  contact:
    name: Plinth Team
    url: https://github.com/jesseiancary/plinth
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://localhost
    description: Development
  - url: https://api.plinth.dev
    description: Production
paths:
...
  /api/v1/auth/register:
    post:
      summary: Register new user
      description: Create a new user account and personal organization
      operationId: register
      tags:
        - Auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - name
              properties:
                email:
                  type: string
                  format: email
                  description: User email address
                password:
                  type: string
                  format: password
                  minLength: 8
                  maxLength: 128
                  description: User password (min 8 characters)
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
                  description: User full name
            examples:
              newUser:
                summary: Register new user
                value:
                  email: sarah.chen@example.com
                  password: 'MyS3cur3P@ssw0rd!'
                  name: Sarah Chen
      responses:
        '201':
          description: User created successfully
          headers:
            Set-Cookie:
              schema:
                type: string
                example: refreshToken=...; HttpOnly; Secure; SameSite=Strict
          content:
            application/json:
              schema:
                type: object
                required:
                  - accessToken
                  - user
                properties:
                  accessToken:
                    type: string
                    description: JWT access token
                  user:
                    $ref: '#/components/schemas/User'
              examples:
                userCreated:
                  summary: User registered successfully
                  value:
                    accessToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiJ1c3JfNGtMOG1OMnBROXZSNndZM3NGN3RYMSIsImlhdCI6MTcxNjg5MzIwMCwiZXhwIjoxNzE2ODk0MTAwfQ.signature
                    user:
                      id: usr_4kL8mN2pQ9vR6wY3sF7tX1
                      email: sarah.chen@example.com
                      name: Sarah Chen
                      createdAt: '2026-05-28T14:32:15.123Z'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                passwordTooShort:
                  summary: Password validation failed
                  value:
                    error:
                      code: VALIDATION_ERROR
                      message: Validation failed
                      details:
                        password: Password must be at least 8 characters
                invalidEmail:
                  summary: Invalid email format
                  value:
                    error:
                      code: VALIDATION_ERROR
                      message: Validation failed
                      details:
                        email: Invalid email format
                registrationFailed:
                  summary: Registration failed (generic error prevents account enumeration)
                  value:
                    error:
                      code: REGISTRATION_FAILED
                      message: Unable to complete registration
                      details: {}
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /api/v1/auth/login:
    post:
      summary: Login user
      description: Authenticate user and return access token
      operationId: login
      tags:
        - Auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  description: User email address
                password:
                  type: string
                  format: password
                  description: User password
            examples:
              validLogin:
                summary: Login with email and password
                value:
                  email: sarah.chen@example.com
                  password: 'MyS3cur3P@ssw0rd!'
      responses:
        '200':
          description: Login successful
          headers:
            Set-Cookie:
              schema:
                type: string
                example: refreshToken=...; HttpOnly; Secure; SameSite=Strict
          content:
            application/json:
              schema:
                type: object
                required:
                  - accessToken
                  - user
                properties:
                  accessToken:
                    type: string
                    description: JWT access token
                  user:
                    $ref: '#/components/schemas/User'
              examples:
                loginSuccess:
                  summary: Login successful
                  value:
                    accessToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiJ1c3JfNGtMOG1OMnBROXZSNndZM3NGN3RYMSIsImlhdCI6MTcxNjg5MzIwMCwiZXhwIjoxNzE2ODk0MTAwfQ.signature
                    user:
                      id: usr_4kL8mN2pQ9vR6wY3sF7tX1
                      email: sarah.chen@example.com
                      name: Sarah Chen
                      createdAt: '2026-05-15T09:23:41.123Z'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalidCredentials:
                  summary: Wrong password or email not found
                  value:
                    error:
                      code: INVALID_CREDENTIALS
                      message: Invalid credentials
                      details: {}
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /api/v1/auth/refresh:
    post:
      summary: Refresh access token
      description: |
        Use the httpOnly refresh token cookie to obtain a new access token.

        The refresh token is automatically sent via cookie - no request body required.

        This endpoint rotates the refresh token (issues a new one and invalidates the old one).

        Call this endpoint before your access token expires (15 minute lifetime).
      operationId: refreshToken
      tags:
        - Auth
      security: []
      responses:
        '200':
          description: Token refreshed successfully
          headers:
            Set-Cookie:
              schema:
                type: string
                example: refreshToken=...; HttpOnly; Secure; SameSite=Strict
          content:
            application/json:
              schema:
                type: object
                required:
                  - accessToken
                properties:
                  accessToken:
                    type: string
                    description: New JWT access token
              examples:
                tokenRefreshed:
                  summary: New access token issued
                  value:
                    accessToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiJ1c3JfNGtMOG1OMnBROXZSNndZM3NGN3RYMSIsImlhdCI6MTcxNjg5NDIwMCwiZXhwIjoxNzE2ODk1MTAwfQ.new_signature
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                missingToken:
                  summary: No refresh token cookie provided
                  value:
                    error:
                      code: REFRESH_TOKEN_MISSING
                      message: Refresh token not found
                      details: {}
                invalidToken:
                  summary: Refresh token signature invalid
                  value:
                    error:
                      code: INVALID_REFRESH_TOKEN
                      message: Invalid refresh token
                      details: {}
                expiredToken:
                  summary: Refresh token has expired (7 day lifetime)
                  value:
                    error:
                      code: REFRESH_TOKEN_EXPIRED
                      message: Refresh token expired
                      details: {}
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
...
  /api/v1/auth/me:
    get:
      summary: Get current user
      description: Get currently authenticated user with their memberships
      operationId: getCurrentUser
      tags:
        - Auth
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user details
          content:
            application/json:
              schema:
                type: object
                required:
                  - id
                  - email
                  - name
                  - createdAt
                  - memberships
                properties:
                  id:
                    type: string
                    description: User ID
                  email:
                    type: string
                    format: email
                    description: User email
                  name:
                    type: string
                    description: User name
                  createdAt:
                    type: string
                    format: date-time
                    description: Account creation timestamp
                  memberships:
                    type: array
                    items:
                      type: object
                      required:
                        - id
                        - role
                        - organization
                        - createdAt
                      properties:
                        id:
                          type: string
                        role:
                          type: string
                          enum: [OWNER, ADMIN, MEMBER]
                        organization:
                          type: object
                          required:
                            - id
                            - name
                            - slug
                          properties:
                            id:
                              type: string
                            name:
                              type: string
                            slug:
                              type: string
                        createdAt:
                          type: string
                          format: date-time
              examples:
                currentUserWithMemberships:
                  summary: User with multiple organizations
                  value:
                    id: usr_4kL8mN2pQ9vR6wY3sF7tX1
                    email: sarah.chen@example.com
                    name: Sarah Chen
                    createdAt: '2026-05-15T09:23:41.123Z'
                    memberships:
                      - id: mem_9zX2pL5qK8vN4wR7tY3sF6
                        role: OWNER
                        organization:
                          id: org_2mK9x3pQ7nL4vB8wR1tY6sF5
                          name: Acme Engineering
                          slug: acme-engineering
                        createdAt: '2026-05-15T09:23:41.123Z'
                      - id: mem_5pL3qK7vN2wR9tY4sF8xB1
                        role: ADMIN
                        organization:
                          id: org_8nM6pQ4vR1wY7sF2tX5kL9
                          name: Beta Labs
                          slug: beta-labs
                        createdAt: '2026-05-20T11:14:52.456Z'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                notAuthenticated:
                  summary: No access token provided or token invalid
                  value:
                    error:
                      code: UNAUTHENTICATED
                      message: Authentication required
                      details: {}
  /api/v1/auth/password:
    patch:
      summary: Change password
      description: |
        Change the user's password. This operation:
        - Validates the current password
        - Enforces strong password policy (8+ chars, uppercase, lowercase, number, special char)
        - Increments tokenVersion to invalidate ALL existing sessions across all devices
        - Clears the refresh token cookie

        **Security Note:** All active sessions will be terminated. User must log in again with the new password.
      operationId: changePassword
      tags:
        - Auth
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - currentPassword
                - newPassword
              properties:
                currentPassword:
                  type: string
                  minLength: 1
                  description: User's current password for verification
                newPassword:
                  type: string
                  minLength: 8
                  maxLength: 128
                  pattern: '^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[^A-Za-z0-9]).{8,}$'
                  description: New password (must contain uppercase, lowercase, number, special character)
            examples:
              passwordChange:
                summary: Change password request
                value:
                  currentPassword: OldP@ssword123
                  newPassword: NewSecureP@ss456
      responses:
        '204':
          description: Password changed successfully, all sessions invalidated
        '400':
          description: Validation error (weak password)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                weakPassword:
                  summary: New password doesn't meet requirements
                  value:
                    error:
                      code: VALIDATION_ERROR
                      message: Validation failed
                      details:
                        issues:
                          - code: invalid_string
                            message: Password must contain at least one uppercase letter
                            path: ['newPassword']
        '401':
          description: Current password is incorrect or not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalidPassword:
                  summary: Current password is incorrect
                  value:
                    error:
                      code: INVALID_PASSWORD
                      message: Current password is incorrect
                      details: {}
                notAuthenticated:
                  summary: No access token provided
                  value:
                    error:
                      code: UNAUTHENTICATED
                      message: Authentication required
                      details: {}
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
...
components:
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
            details:
              type: object
              description: Additional error context
              additionalProperties: true
    User:
      type: object
      required:
        - id
        - email
        - name
        - createdAt
      properties:
        id:
          type: string
          description: User ID
        email:
          type: string
          format: email
          description: User email address
        name:
          type: string
          description: User full name
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
...
  responses:
    RateLimitExceeded:
      description: Rate limit exceeded
      headers:
        RateLimit-Limit:
          schema:
            type: integer
          description: Maximum number of requests allowed in the window
        RateLimit-Remaining:
          schema:
            type: integer
          description: Number of requests remaining in the current window
        RateLimit-Reset:
          schema:
            type: integer
          description: Unix timestamp when the rate limit window resets
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
      content:
        application/json:
          schema:
            type: object
            required:
              - error
            properties:
              error:
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                    enum: [RATE_LIMIT_EXCEEDED]
                  message:
                    type: string
                    description: Human-readable rate limit message
                  details:
                    type: object
          examples:
            authLogin:
              summary: Login rate limit exceeded
              value:
                error:
                  code: RATE_LIMIT_EXCEEDED
                  message: Too many login attempts. Please try again in 1 minute.
            authRegister:
              summary: Registration rate limit exceeded
              value:
                error:
                  code: RATE_LIMIT_EXCEEDED
                  message: Too many registration attempts. Please try again in 1 hour.
            invitationCreate:
              summary: Invitation creation rate limit exceeded
              value:
                error:
                  code: RATE_LIMIT_EXCEEDED
                  message: Invitation limit reached. Please try again later.
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token

    apiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: API key (format sk_live_...)
security:
  - bearerAuth: []
  - apiKeyAuth: []
...